WIP on moving logic in discrete steps

+++ shared/src/main/scala/shared/model/Position.scala
@@ -1,3 +1,4 @@
 package shared.model
 
-case class Position(x: Int, y: Int)
+case class Position(x: Double, y: Double)
+
